split out ordered_tree.go
